pyproject.toml
[project]
name = "flex-evals-viewer"
version = "0.1.0"
description = "Template repository showing how to use flex-evals to evaluate LLM-based agents"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flex-evals>=0.3.26",
    "sik-llms>=0.3.30",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["evals"]
timeout = 120

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-timeout>=2.4.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "ruff>=0.15.2",
]
